More refactor between two discussions pages
